Refactoring some duplication out and making clippy happy

+++ tests/domains_push_test.rs
@@ -10,7 +10,7 @@ fn test_initiate_push_test() {
         "POST",
     );
     let client = setup.0;
-    let account_id = 1385 as u64;
+    let account_id = 1385_u64;
     let domain = "target-account.test";
     let payload = InitiatePushPayload {
         new_account_email: String::from("admin@target-account.test"),
@@ -36,7 +36,7 @@ fn test_initiate_push_test() {
 fn test_list_pushes() {
     let setup = setup_mock_for("/1385/domains/pushes", "listPushes/success", "GET");
     let client = setup.0;
-    let account_id = 1385 as u64;
+    let account_id = 1385_u64;
 
     let response = client.domains().list_pushes(account_id, None).unwrap();
     let domain_pushes_list = response.data.unwrap();
@@ -48,7 +48,7 @@ fn test_list_pushes() {
 fn test_accept_push() {
     let setup = setup_mock_for("/1385/domains/pushes/42", "acceptPush/success", "POST");
     let client = setup.0;
-    let account_id = 1385 as u64;
+    let account_id = 1385_u64;
     let push_id = 42;
 
     let response = client.domains().accept_push(account_id, push_id);
@@ -60,7 +60,7 @@ fn test_accept_push() {
 fn test_reject_push() {
     let setup = setup_mock_for("/1385/domains/pushes/42", "rejectPush/success", "DELETE");
     let client = setup.0;
-    let account_id = 1385 as u64;
+    let account_id = 1385_u64;
     let push_id = 42;
 
     let response = client.domains().reject_push(account_id, push_id);
